pyproject.toml
[tool.poetry]
name = "helloworld-django"
version = "0.1.0"
description = ""
authors = ["Mukagali <mukgaly.omirzak2003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-mptt = "^0.15.0"
django-cors-headers = "^4.3.1"
django-filter = "^23.4"
dj-rest-auth = "^5.0.2"
djangorestframework-simplejwt = "^5.3.0"
django-timedeltafield = "^0.7.10"
django-rest-swagger = "^2.2.0"
drf-spectacular = "^0.26.5"
djoser = "^2.2.2"
django-storages = "^1.14.2"
pillow = "^10.1.0"
boto3 = "^1.34.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
